+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2 - 16/01/21
+- Added initialisation of environment tiles, along with addition of energy sources and obstacles
+- Added visualisation function of environment tiles
+- Moved stuff to 'src' subdirectory
+
 ## 0.1.1 - 15/01/21
 - Added to-do list
 - Changed repo name and local project name from 'CPP-Project' to 'Agent_Survival_Simulation'